Upgraded cryptocompare to v3 framework
