+++ .changeset/selfish-schools-doubt.md
@@ -1,0 +1,5 @@
+---
+'@chainlink/cryptocompare-adapter': major
+---
+
+Upgraded adapter from v2 to v3 framework
