+++ packages/core/legos/tsconfig.json
@@ -155,9 +155,6 @@
       "path": "../../sources/cryptocompare"
     },
     {
-      "path": "../../sources/cryptocompare-test"
-    },
-    {
       "path": "../../sources/cryptoid"
     },
     {
