--- packages/sources/cryptocompare-test/CHANGELOG.md
@@ -1,73 +1,0 @@
-# @chainlink/cryptocompare-test-adapter
-
-## 1.3.1
-
-### Patch Changes
-
-- 77ad946a9: Bumped framework version
-
-## 1.3.0
-
-### Minor Changes
-
-- 9ae05bdd2: Added custom routing using WS_ENABLED env var
-
-## 1.2.6
-
-### Patch Changes
-
-- 4245b4d8b: Bumped framework version
-
-## 1.2.5
-
-### Patch Changes
-
-- 8fe5be225: Bumped framework version
-
-## 1.2.4
-
-### Patch Changes
-
-- c9cee149e: Bump fw version
-
-## 1.2.3
-
-### Patch Changes
-
-- 2fdaa5aa4: Bump v3 version
-
-## 1.2.2
-
-### Patch Changes
-
-- bc8886c63: Bumped framework version
-
-## 1.2.1
-
-### Patch Changes
-
-- 37c0d8883: Bumped v3 framework version
-
-## 1.2.0
-
-### Minor Changes
-
-- d474a5553: Update v3 framework to latest version
-
-## 1.1.1
-
-### Patch Changes
-
-- 65014014d: Bumped v3 framework version. Upgraded typescript to version 4.9.5
-
-## 1.1.0
-
-### Minor Changes
-
-- 8b3c35545: Update ea-framework version and transports, add vwap, marketcap, volume endpoints
-
-## 1.0.0
-
-### Major Changes
-
-- cf8609ad5: Added Cryptocompare-test adapter
