--- packages/sources/cryptocompare-test/src/endpoint/index.ts
@@ -1,4 +1,0 @@
-export { endpoint as crypto } from './crypto-router'
-export { endpoint as vwap } from './vwap'
-export { endpoint as marketcap } from './marketcap'
-export { endpoint as volume } from './volume'
