--- packages/sources/cryptocompare/schemas/env.json
@@ -1,18 +1,0 @@
-{
-  "$id": "https://external-adapters.chainlinklabs.com/schemas/cryptocompare-adapter.json",
-  "title": "Chainlink External Adapter for CryptoCompare",
-  "required": ["API_KEY"],
-  "type": "object",
-  "properties": {
-    "API_KEY": {
-      "type": "string",
-      "description": "An API key that can be obtained from [here](https://min-api.cryptocompare.com/pricing)",
-      "required": false
-    }
-  },
-  "allOf": [
-    {
-      "$ref": "https://external-adapters.chainlinklabs.com/schemas/ea-bootstrap.json"
-    }
-  ]
-}
