--- packages/sources/cryptocompare/src/config/limits.json
@@ -1,42 +1,0 @@
-{
-  "http": {
-    "free": {
-      "rateLimit1h": 136.98,
-      "rateLimit1mo": 100000,
-      "note": "only mentions monthly limit"
-    },
-    "professional": {
-      "rateLimit1h": 342.46,
-      "rateLimit1mo": 250000,
-      "note": "only mentions monthly limit"
-    },
-    "corporate": {
-      "rateLimit1h": 1027.39,
-      "rateLimit1mo": 750000,
-      "note": "only mentions monthly limit"
-    },
-    "enterprise-lite": {
-      "rateLimit1h": 2083,
-      "rateLimit1mo": 1500000
-    }
-  },
-  "ws": {
-    "free": {
-      "connections": -1,
-      "tickerSupport": false,
-      "subscriptions": 50
-    },
-    "professional": {
-      "connections": -1,
-      "subscriptions": 500
-    },
-    "corporate": {
-      "connections": -1,
-      "subscriptions": 500
-    },
-    "enterprise-lite": {
-      "connections": -1,
-      "subscriptions": 500
-    }
-  }
-}
