+++ packages/sources/cryptocompare/src/config/overrides.json
