--- packages/sources/cryptocompare/src/config/symbols.json
@@ -1,5 +1,0 @@
-{
-  "cryptocompare": {
-    "PAX": "USDP"
-  }
-}
