+++ packages/sources/cryptocompare/src/endpoint/index.ts
@@ -1,7 +1,4 @@
-import type { TInputParameters as CryptoInputParameters } from './crypto'
-import type { TInputParameters as VwapInputParameters } from './vwap'
-
-export type TInputParameters = CryptoInputParameters | VwapInputParameters
-
-export * as crypto from './crypto'
-export * as vwap from './vwap'
+export { endpoint as crypto } from './crypto-router'
+export { endpoint as vwap } from './vwap'
+export { endpoint as marketcap } from './marketcap'
+export { endpoint as volume } from './volume'
