+++ packages/sources/cryptocompare/src/index.ts
@@ -1,10 +1,29 @@
-import { expose } from '@chainlink/ea-bootstrap'
-import { makeExecute, endpointSelector, makeWSHandler } from './adapter'
-import * as endpoints from './endpoint'
-import { makeConfig, NAME } from './config'
-import rateLimit from './config/limits.json'
+import { expose, ServerInstance } from '@chainlink/external-adapter-framework'
+import { PriceAdapter } from '@chainlink/external-adapter-framework/adapter'
+import { crypto, vwap, volume, marketcap } from './endpoint'
+import { config } from './config'
 
-const adapterContext = { name: NAME, rateLimit }
+export const adapter = new PriceAdapter({
+  name: 'CRYPTOCOMPARE',
+  defaultEndpoint: 'crypto',
+  config,
+  endpoints: [crypto, vwap, volume, marketcap],
+  rateLimiting: {
+    tiers: {
+      free: {
+        rateLimit1h: 136.98,
+      },
+      professional: {
+        rateLimit1h: 342.46,
+      },
+      corporate: {
+        rateLimit1h: 1027.39,
+      },
+      'enterprise-lite': {
+        rateLimit1h: 2083,
+      },
+    },
+  },
+})
 
-const { server } = expose(adapterContext, makeExecute(), makeWSHandler(), endpointSelector)
-export { NAME, makeExecute, makeConfig, server, endpoints }
+export const server = (): Promise<ServerInstance | undefined> => expose(adapter)
