+++ packages/sources/cryptocompare/test/integration/fixtures.ts
@@ -1,132 +1,155 @@
 import nock from 'nock'
 
-export const mockPriceResponseSuccess = (): nock.Scope =>
-  nock('https://min-api.cryptocompare.com/', {
+export const mockCryptoSuccess = (): nock.Scope =>
+  nock('https://min-api.cryptocompare.com', {
     encodedQueryParams: true,
-    reqheaders: {
-      authorization: 'Apikey fake-api-key',
-    },
   })
     .get('/data/pricemultifull')
-    .query({ fsyms: 'ETH', tsyms: 'BTC' })
+    .query({ fsyms: 'ETH', tsyms: 'USD' })
     .reply(
       200,
-      () => ({
+      {
         RAW: {
           ETH: {
-            BTC: {
+            USD: {
               TYPE: '5',
               MARKET: 'CCCAGG',
               FROMSYMBOL: 'ETH',
-              TOSYMBOL: 'BTC',
+              TOSYMBOL: 'USD',
               FLAGS: '2052',
-              PRICE: 0.06543,
-              LASTUPDATE: 1634917845,
-              MEDIAN: 0.06541,
-              LASTVOLUME: 0.3667,
-              LASTVOLUMETO: 0.0239964813,
-              LASTTRADEID: '266555634',
-              VOLUMEDAY: 426154.9210721772,
-              VOLUMEDAYTO: 44337.6725074485,
-              VOLUME24HOUR: 199268.35105042686,
-              VOLUME24HOURTO: 12999.111610148046,
-              OPENDAY: 0.06516,
-              HIGHDAY: 0.06595,
-              LOWDAY: 0.06486,
-              OPEN24HOUR: 0.06492,
-              HIGH24HOUR: 0.06595,
-              LOW24HOUR: 0.06438,
-              LASTMARKET: 'Binance',
-              VOLUMEHOUR: 21826.54200988967,
-              VOLUMEHOURTO: 1662.8001349818653,
-              OPENHOUR: 0.06539,
-              HIGHHOUR: 0.06563,
-              LOWHOUR: 0.06531,
-              TOPTIERVOLUME24HOUR: 196133.9280134269,
-              TOPTIERVOLUME24HOURTO: 12794.391640956886,
-              CHANGE24HOUR: 0.0005099999999999966,
-              CHANGEPCT24HOUR: 0.7855822550831739,
-              CHANGEDAY: 0.00027000000000000635,
-              CHANGEPCTDAY: 0.41436464088398767,
-              CHANGEHOUR: 0.00003999999999999837,
-              CHANGEPCTHOUR: 0.06117143294081415,
+              PRICE: 1276.75,
+              LASTUPDATE: 1666357726,
+              MEDIAN: 1276.59,
+              LASTVOLUME: 0.6333,
+              LASTVOLUMETO: 808.876092,
+              LASTTRADEID: '34667055',
+              VOLUMEDAY: 287782.6125126447,
+              VOLUMEDAYTO: 367505833.00910026,
+              VOLUME24HOUR: 573086.14879595,
+              VOLUME24HOURTO: 735584356.0525743,
+              OPENDAY: 1282.68,
+              HIGHDAY: 1292.33,
+              LOWDAY: 1254.6,
+              OPEN24HOUR: 1298.03,
+              HIGH24HOUR: 1311.05,
+              LOW24HOUR: 1253.35,
+              LASTMARKET: 'binanceusa',
+              VOLUMEHOUR: 22103.92288265006,
+              VOLUMEHOURTO: 28158189.562133063,
+              OPENHOUR: 1269.67,
+              HIGHHOUR: 1277.56,
+              LOWHOUR: 1268.24,
+              TOPTIERVOLUME24HOUR: 573086.14879595,
+              TOPTIERVOLUME24HOURTO: 735584356.0525743,
+              CHANGE24HOUR: -21.279999999999973,
+              CHANGEPCT24HOUR: -1.6394074096900666,
+              CHANGEDAY: -5.930000000000064,
+              CHANGEPCTDAY: -0.4623132815667246,
+              CHANGEHOUR: 7.079999999999927,
+              CHANGEPCTHOUR: 0.5576252097001525,
               CONVERSIONTYPE: 'direct',
               CONVERSIONSYMBOL: '',
-              SUPPLY: 117601698.8115,
-              MKTCAP: 7694679.153236445,
+              SUPPLY: 122373863.499,
+              MKTCAP: 156240830222.34824,
               MKTCAPPENALTY: 0,
-              CIRCULATINGSUPPLY: 117601698.8115,
-              CIRCULATINGSUPPLYMKTCAP: 7694679.153236445,
-              TOTALVOLUME24H: 2239758.9103871463,
-              TOTALVOLUME24HTO: 146508.4089075496,
-              TOTALTOPTIERVOLUME24H: 2231978.1419115765,
-              TOTALTOPTIERVOLUME24HTO: 145999.6785563128,
+              CIRCULATINGSUPPLY: 122373863.499,
+              CIRCULATINGSUPPLYMKTCAP: 156240830222.34824,
+              TOTALVOLUME24H: 2594280.295149976,
+              TOTALVOLUME24HTO: 3316143982.4100766,
+              TOTALTOPTIERVOLUME24H: 2584883.3953427817,
+              TOTALTOPTIERVOLUME24HTO: 3304146490.5812416,
               IMAGEURL: '/media/37746238/eth.png',
             },
           },
         },
         DISPLAY: {
           ETH: {
-            BTC: {
+            USD: {
               FROMSYMBOL: 'Ξ',
-              TOSYMBOL: 'Ƀ',
+              TOSYMBOL: '$',
               MARKET: 'CryptoCompare Index',
-              PRICE: 'Ƀ 0.06543',
+              PRICE: '$ 1,276.75',
               LASTUPDATE: 'Just now',
-              LASTVOLUME: 'Ξ 0.3667',
-              LASTVOLUMETO: 'Ƀ 0.02400',
-              LASTTRADEID: '266555634',
-              VOLUMEDAY: 'Ξ 426,154.9',
-              VOLUMEDAYTO: 'Ƀ 44,337.7',
-              VOLUME24HOUR: 'Ξ 199,268.4',
-              VOLUME24HOURTO: 'Ƀ 12,999.1',
-              OPENDAY: 'Ƀ 0.06516',
-              HIGHDAY: 'Ƀ 0.06595',
-              LOWDAY: 'Ƀ 0.06486',
-              OPEN24HOUR: 'Ƀ 0.06492',
-              HIGH24HOUR: 'Ƀ 0.06595',
-              LOW24HOUR: 'Ƀ 0.06438',
-              LASTMARKET: 'Binance',
-              VOLUMEHOUR: 'Ξ 21,826.5',
-              VOLUMEHOURTO: 'Ƀ 1,662.80',
-              OPENHOUR: 'Ƀ 0.06539',
-              HIGHHOUR: 'Ƀ 0.06563',
-              LOWHOUR: 'Ƀ 0.06531',
-              TOPTIERVOLUME24HOUR: 'Ξ 196,133.9',
-              TOPTIERVOLUME24HOURTO: 'Ƀ 12,794.4',
-              CHANGE24HOUR: 'Ƀ 0.00051',
-              CHANGEPCT24HOUR: '0.79',
-              CHANGEDAY: 'Ƀ 0.00027',
-              CHANGEPCTDAY: '0.41',
-              CHANGEHOUR: 'Ƀ 0.000040',
-              CHANGEPCTHOUR: '0.06',
+              LASTVOLUME: 'Ξ 0.6333',
+              LASTVOLUMETO: '$ 808.88',
+              LASTTRADEID: '34667055',
+              VOLUMEDAY: 'Ξ 287,782.6',
+              VOLUMEDAYTO: '$ 367,505,833.0',
+              VOLUME24HOUR: 'Ξ 573,086.1',
+              VOLUME24HOURTO: '$ 735,584,356.1',
+              OPENDAY: '$ 1,282.68',
+              HIGHDAY: '$ 1,292.33',
+              LOWDAY: '$ 1,254.60',
+              OPEN24HOUR: '$ 1,298.03',
+              HIGH24HOUR: '$ 1,311.05',
+              LOW24HOUR: '$ 1,253.35',
+              LASTMARKET: 'binanceusa',
+              VOLUMEHOUR: 'Ξ 22,103.9',
+              VOLUMEHOURTO: '$ 28,158,189.6',
+              OPENHOUR: '$ 1,269.67',
+              HIGHHOUR: '$ 1,277.56',
+              LOWHOUR: '$ 1,268.24',
+              TOPTIERVOLUME24HOUR: 'Ξ 573,086.1',
+              TOPTIERVOLUME24HOURTO: '$ 735,584,356.1',
+              CHANGE24HOUR: '$ -21.28',
+              CHANGEPCT24HOUR: '-1.64',
+              CHANGEDAY: '$ -5.93',
+              CHANGEPCTDAY: '-0.46',
+              CHANGEHOUR: '$ 7.08',
+              CHANGEPCTHOUR: '0.56',
               CONVERSIONTYPE: 'direct',
               CONVERSIONSYMBOL: '',
-              SUPPLY: 'Ξ 117,601,698.8',
-              MKTCAP: 'Ƀ 7.69 M',
+              SUPPLY: 'Ξ 122,373,863.5',
+              MKTCAP: '$ 156.24 B',
               MKTCAPPENALTY: '0 %',
-              CIRCULATINGSUPPLY: 'Ξ 117,601,698.8',
-              CIRCULATINGSUPPLYMKTCAP: 'Ƀ 7.69 M',
-              TOTALVOLUME24H: 'Ξ 2.24 M',
-              TOTALVOLUME24HTO: 'Ƀ 146.51 K',
-              TOTALTOPTIERVOLUME24H: 'Ξ 2.23 M',
-              TOTALTOPTIERVOLUME24HTO: 'Ƀ 146.00 K',
+              CIRCULATINGSUPPLY: 'Ξ 122,373,863.5',
+              CIRCULATINGSUPPLYMKTCAP: '$ 156.24 B',
+              TOTALVOLUME24H: 'Ξ 2.59 M',
+              TOTALVOLUME24HTO: '$ 3.32 B',
+              TOTALTOPTIERVOLUME24H: 'Ξ 2.58 M',
+              TOTALTOPTIERVOLUME24HTO: '$ 3.30 B',
               IMAGEURL: '/media/37746238/eth.png',
             },
           },
         },
-      }),
+      },
       [
+        'Server',
+        'nginx',
+        'Date',
+        'Fri, 21 Oct 2022 13:08:55 GMT',
         'Content-Type',
-        'application/json',
+        'application/json; charset=UTF-8',
+        'Transfer-Encoding',
+        'chunked',
         'Connection',
         'close',
         'Vary',
         'Accept-Encoding',
-        'Vary',
-        'Origin',
+        'Content-Security-Policy',
+        "frame-ancestors 'none'",
+        'Access-Control-Allow-Origin',
+        '*',
+        'Access-Control-Allow-Methods',
+        'GET, POST, OPTIONS',
+        'Access-Control-Allow-Headers',
+        'Content-Type, Cookie, Set-Cookie, Authorization',
+        'Access-Control-Allow-Credentials',
+        'true',
+        'Cache-Control',
+        'public, max-age=10',
+        'X-CryptoCompare-Cache-HIT',
+        'true',
+        'X-CryptoCompare-Server-Id',
+        'ccc-api12',
       ],
     )
+    .persist()
+
+export const mockVwapSuccess = (): nock.Scope =>
+  nock('https://min-api.cryptocompare.com', {
+    encodedQueryParams: true,
+  })
     .get('/data/dayAvg')
     .query({ fsym: 'AMPL', tsym: 'USD', toTs: /.+/i })
     .reply(200, { USD: 0.9224, ConversionType: { type: 'direct', conversionSymbol: '' } }, [
@@ -139,92 +162,3 @@ export const mockPriceResponseSuccess = (): nock.Scope =>
       'Vary',
       'Origin',
     ])
-
-export const mockPriceResponseFailure = (): nock.Scope =>
-  nock('https://min-api.cryptocompare.com/', {
-    encodedQueryParams: true,
-    reqheaders: {
-      authorization: 'Apikey fake-api-key',
-    },
-  })
-    .get('/data/pricemultifull')
-    .query({ fsyms: 'ETH', tsyms: 'XXX' })
-    .reply(500, () => 'Invalid base or quote', [
-      'Content-Type',
-      'application/json',
-      'Connection',
-      'close',
-      'Vary',
-      'Accept-Encoding',
-      'Vary',
-      'Origin',
-    ])
-
-export const mockSubscribeResponse = {
-  request: {
-    action: 'SubAdd',
-    subs: ['5~CCCAGG~ETH~BTC'],
-  },
-  response: [
-    {
-      TYPE: '16',
-      MESSAGE: 'SUBSCRIBECOMPLETE',
-      SUB: '5~CCCAGG~ETH~BTC',
-    },
-    {
-      TYPE: '3',
-      MESSAGE: 'LOADCOMPLETE',
-      INFO: 'All your valid subs have been loaded.',
-    },
-    {
-      TYPE: '5',
-      MARKET: 'CCCAGG',
-      FROMSYMBOL: 'ETH',
-      TOSYMBOL: 'BTC',
-      FLAGS: 2,
-      PRICE: 0.07038,
-      LASTUPDATE: 1645027932,
-      MEDIAN: 0.07038,
-      LASTVOLUME: 0.045,
-      LASTVOLUMETO: 0.00316565955,
-      LASTTRADEID: '16450279320001',
-      VOLUMEDAY: 253257.38626818382,
-      VOLUMEDAYTO: 11398.462174583636,
-      VOLUME24HOUR: 78687.65999060342,
-      VOLUME24HOURTO: 5583.496196776957,
-      OPENDAY: 0.07148,
-      HIGHDAY: 0.07187,
-      LOWDAY: 0.07014,
-      OPEN24HOUR: 0.07057,
-      HIGH24HOUR: 0.0719,
-      LOW24HOUR: 0.07013,
-      LASTMARKET: 'aax',
-      VOLUMEHOUR: 9845.973204030126,
-      VOLUMEHOURTO: 406.67115535224315,
-      OPENHOUR: 0.07033,
-      HIGHHOUR: 0.07043,
-      LOWHOUR: 0.07021,
-      TOPTIERVOLUME24HOUR: 78687.39639060342,
-      TOPTIERVOLUME24HOURTO: 5583.4778180780095,
-      CURRENTSUPPLY: 119524383.0615,
-      CIRCULATINGSUPPLY: 119524383.0615,
-      MAXSUPPLY: -1,
-      MKTCAPPENALTY: 0,
-      CURRENTSUPPLYMKTCAP: 8412126.079868369,
-      CIRCULATINGSUPPLYMKTCAP: 8412126.079868369,
-      MAXSUPPLYMKTCAP: -1,
-    },
-  ],
-}
-
-export const mockUnsubscribeResponse = {
-  request: {
-    action: 'SubRemove',
-    subs: ['5~CCCAGG~ETH~BTC'],
-  },
-  response: {
-    TYPE: '17',
-    MESSAGE: 'UNSUBSCRIBECOMPLETE',
-    SUB: '5~CCCAGG~ETH~BTC',
-  },
-}
