+++ packages/tsconfig.json
@@ -267,9 +267,6 @@
       "path": "./sources/cryptocompare"
     },
     {
-      "path": "./sources/cryptocompare-test"
-    },
-    {
       "path": "./sources/cryptoid"
     },
     {
