+++ packages/tsconfig.test.json
@@ -267,9 +267,6 @@
       "path": "./sources/cryptocompare/tsconfig.test.json"
     },
     {
-      "path": "./sources/cryptocompare-test/tsconfig.test.json"
-    },
-    {
       "path": "./sources/cryptoid/tsconfig.test.json"
     },
     {
